travel-agent/requirements.txt
chainlit>=2.9.2
langchain>=1.1.0
langgraph>=1.0.4
langchain-core>=0.3.0
langchain-openai>=0.2.0
typing-extensions>=4.8.0
langgraph-checkpoint>=3.0.1
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0


